+++ src/pages/public_registration/log/LogEntry.tsx
@@ -9,13 +9,13 @@ import UnpublishedOutlinedIcon from '@mui/icons-material/UnpublishedOutlined';
 import { Avatar, Box, Divider, styled, SvgIconProps, Tooltip, Typography } from '@mui/material';
 import { TFunction } from 'i18next';
 import { useTranslation } from 'react-i18next';
-import { LogDateItem } from '../../../components/Log/LogDateItem';
-import { LogMessageAccordion } from '../../../components/Log/LogMessageAccordion';
 import { FileType } from '../../../types/associatedArtifact.types';
 import { ImportSourceLogData, LogEntryObject } from '../../../types/log.types';
 import { Message } from '../../../types/publication_types/ticket.types';
 import { getInitials } from '../../../utils/general-helpers';
 import { getFullName } from '../../../utils/user-helpers';
+import { LogDateItem } from './LogDateItem';
+import { LogMessageAccordion } from './LogMessageAccordion';
 
 const logIconProps: SvgIconProps = { color: 'primary', fontSize: 'small' };
